Cambios 17.1.26 11.42

+++ DTO/SolicitudDeMateriales/LotePliego/UpdateLotePliegoDto.cs
@@ -4,9 +4,9 @@
     {
         public int? idSolicitud { get; set; }
 
-        public int? largo { get; set; }
+        public decimal? largo { get; set; }
 
-        public int? ancho { get; set; }
+        public decimal? ancho { get; set; }
 
         public int? cantidadPliegos { get; set; }
